add: AttendanceContent
